requirements.txt
# RAG系统依赖包
typer>=0.9.0
rich>=13.4.2
langchain>=0.0.267
openai>=0.27.8
faiss-cpu>=1.7.4
python-dotenv>=1.0.0
pandas>=2.0.3
numpy>=1.24.3
scipy>=1.10.1
pyPDF2>=3.0.1
docx2txt>=0.8
pytesseract>=0.3.10
pillow>=9.5.0
sentence-transformers>=2.2.2
chromadb>=0.4.6
tqdm>=4.65.0